fix for issue #68
git-svn-id: http://tradelink.googlecode.com/svn/trunk@666 ccc6a032-7b47-0410-9fd8-1f29c74ce1d2

+++ BrokerServers/AnvilServer/TLAnvil.cpp
@@ -438,8 +438,8 @@ int PosList(CString req)
     while(pos = B_GetNextPosition(iterator))
     {
 		TradeLinkServer::TLPosition p;
-		p.AvgPrice = pos->GetAverageExecPrice().GetMoneyValueForServer()/1024;
-		p.ClosedPL = pos->GetClosedPnl().GetMoneyValueForServer()/1024;
+		p.AvgPrice = pos->GetAverageExecPrice().GetMoneyValueForServer()/(double)1024;
+		p.ClosedPL = pos->GetClosedPnl().GetMoneyValueForServer()/(double)1024;
 		p.Size = pos->GetSize();
 		p.Symbol = CString(pos->GetSymbol());
 		CString msg = p.Serialize();
